Move syntax things from Dhall.Core to a new module Dhall.Syntax
This partially addresses #1127.
