+++ dhall/src/Dhall/Pretty/Internal.hs
@@ -59,8 +59,6 @@ module Dhall.Pretty.Internal (
     , rparen
     ) where
 
-import Dhall.Core
-
 #if MIN_VERSION_base(4,8,0)
 #else
 import Control.Applicative (Applicative(..), (<$>))
@@ -72,6 +70,7 @@ import Data.Text.Prettyprint.Doc (Doc, Pretty, space)
 import Dhall.Map (Map)
 import Dhall.Set (Set)
 import Dhall.Src (Src(..))
+import Dhall.Syntax
 import Numeric.Natural (Natural)
 import Prelude hiding (succ)
 import qualified Data.Text.Prettyprint.Doc.Render.Terminal as Terminal
