+++ dhall/src/Dhall/Pretty/Internal.hs-boot
@@ -1,9 +1,8 @@
 module Dhall.Pretty.Internal where
 
-import Data.Text (Text)
 import Data.Text.Prettyprint.Doc (Pretty, Doc)
 
-import {-# SOURCE #-} Dhall.Core
+import {-# SOURCE #-} Dhall.Syntax
 
 data Ann
 
@@ -12,7 +11,3 @@ prettyVar :: Var -> Doc Ann
 prettyConst :: Const -> Doc Ann
 
 prettyExpr :: Pretty a => Expr s a -> Doc Ann
-
-pretty_ :: Pretty a => a -> Text
-
-escapeText_ :: Text -> Text
