+++ dhall/src/Dhall/Syntax.hs-boot
@@ -1,0 +1,7 @@
+module Dhall.Syntax where
+
+data Var
+
+data Const
+
+data Expr s a
